feat(installer): add an exception for non 'trusted-users'
Summary: When a user is part of 'trusted-users', then they can use any binary
cache they want, including ours. But when they aren't, there's still a case
they can use ours:

- Our public key is in 'trusted-public-keys'
- Our URL is in the '{trusted-,}substituters' field

So, we can check that and special case it.

Also warn about the possibility of key/url drift in the future, for now.

Also add a tagref for the cache url settings so that we can isolate and find
them easily in the future.

Test Plan: Short-circuit the conditionals into these code paths and test them
out a bit.