fix(buck): replace and overhaul buck2 update.sh script
Summary: Now the installer is much more robust as it doesn't use nix-update, it
was fragile in various ways.

On top of that, we now also update the rust toolchain based on the
'rust-toolchain' file; only nightly is handled for now, of course.

Test Plan: `$(sl root)/buck/nix/buck2/update.sh`