chore(buck): wire up platform information
Summary: This wires up the default platform configuration into `.buckconfig` so
that it's propagated everywhere by default. Doing this required looking at the
buck source code a bit, but seems to work.

With this, by default the host platform system is detected and used as a
configuration value located at `prelude//platforms:default`; and then used as
the configuration target for all subsequent builds by default. Because of this,
`select()` is now available, so we can choose different dependencies or targets
based on this information.

Test Plan: `buck run src/bash/echo-test:testing`, which now combines `alias()`
with `select()` to choose a default target.