chore(nix): buck2 unstable-2023-01-20 -> unstable-2023-02-08
Summary: This requires an (extra) patch to buck2 to revert a recent commit
for boxing protobuf fields, used for buck2 events; that functionality requires
subsequent patches to upstream 'prost' and 'tonic' which, currently, Nix can't
handle gracefully (prost uses a multi-package workspace, which causes some failure
in the upstream buildRustPackage function for Nix)

Unfortunately, that patch is a good performance and memory improvement -- but
since we were already taking the hit here, reverting this and coming back to it
later is probably preferable to not being able to update 'buck2' ever again in
the mean time.

Once these patches go upstream (in some form, ideally soon) then we can remove
this.