refactor(buck/prelude): migrate rust to toolchain rules
Summary: Rules providing toolchains are expected to have `is_toolchain_rule`
set to `True`. They should also yield a Provider for that toolchain, and then
rules can depend on the toolchain via toolchain_dep, with an expectation that
the Provider will be available.

This only migrates the Rust toolchain, as an experiment. Others will follow
(once I understand all the semantics and APIs better.)