chore(buck/prelude): large refactor
Summary: This refactors the internal prelude quite a bit, adds a lot of
documentation, a `filegroup()` implementation, but most importantly it
establishes a basic design for implementing toolchain rules and using
`toolchain_dep()` on them.

Much like configuration rules, buck2 expects rules that build toolchains to
have `is_toolchain_rule` set. This is used in combination with the platform
and configuration data to execute remote rules properly, among others. Users
of the toolchain can reasonably depend on this with `attrs.toolchain_dep()` and
establish what providers they need from that rule.

Toolchains are provisioned by Nix, but it isn't clear to me every `nix.build()`
call should result in a toolchain rule. Also, rules can't be created after
analysis, which is when this needs to be established. Finally, a rule for a
toolchain needs to provide its own providers as well as the `DefaultInfo()`,
so abstracting it at the rule level is difficult. So this refactors the
`@prelude//nixpkgs.bzl` interface to both export `nix.build()` calls as well as
export macros/functions that other rules can use to provision toolchains with
`is_toolchain_rule` set properly. But to re-use the rule logic, we also have to
re-use the attribute logic. A bit of a chore.

In the future, if it looks like `nix.build()` should always result in a
toolchain, we can clean this up a bit. But the need to return providers as well
as the analysis time restrictions mean it probably won't change substantially.

The first toolchain to be migrated to explicit `toolchain_dep()` is Rust.
Others can follow as needed.