feat(src): basic renode examples
This enables some basic fun with Renode. Unfortunately, it can't be used in a
turnkey manner yet with a test target, since the testing integration with Robot
Framework is broken (see the TARGETS file for comments). But it does work! The
following will launch a working Renode instance of a HiFive Unmatched, booting
a Linux image:

    buck run src/renode:hifive-unleashed