fix: remove installer hack that was fixed in nix-2.14.0
Summary: The Nix 2.14.0 release notes have an entry:

    In flakes, the .outPath attribute of a flake now always refers to the
	directory containing the flake.nix. This was not the case for when
	flake.nix was in a subdirectory of e.g. a Git repository. The root of
	the source of a flake in a subdirectory is still available in
	.sourceInfo.outPath.

We were working around this in the installer with a gross hack, to `cd` into
the right place for buildRustPackage. So it's was completely broken on Nix
2.14.0 or later. Delete the hack, upgrade the CI builders, and set the MSNV to
2.14.0+ (good thing I added this earlier!)

This should fix the issue, though the installer needs to be tested with 'nix
run' still.

Related Issue: #8

Test Plan: Just go for it.